tests/requirements_test.txt
pytest==7.4.0
pytest-flask==1.2.0
psycopg2-binary==2.9.9
